Fix Whitaker Dylint suite findings across the workspace
Prepare the workspace for the full Whitaker Dylint suite by fixing
every finding it raises, without altering any public API surface:

- no_expect_outside_tests: test helpers and fixtures no longer call
  `.expect(...)`; they return `Result` and propagate errors, or use
  `let ... else { panic!(...) }` / match-with-panic for invariants.
  The `demo_input` rstest fixture becomes fallible and its consumers
  apply `?`.
- no_unwrap_or_else_panic: panicking `unwrap_or_else` closures in
  test helpers become let-else panics.
- bumpy_road_function: nested conditional clusters are extracted
  into named helpers (`FluentLocalizerBuilder::try_build` bundle
  construction, `apply_field_attr` boolean parsing, PowerShell
  integer/float type mapping, bridge manifest dependency rendering,
  the hello_world command dispatch, and the extends-scenario jail
  setup).
- module_must_have_inner_docs: inline test and support modules open
  with a `//!` summary.
- module_max_lines: oversized modules are split into sibling
  submodules (docs field resolution, ortho-attrs unknown-key tests,
  and the agent-context test builders).

Behaviour is unchanged; the split modules re-export or import the
moved items so all call sites compile as before.
