+++ ortho_config/src/agent_context/tests.rs
@@ -247,16 +247,17 @@ fn mutation_effect_serializes_canonical_wire_values(
 }
 
 fn field<'a>(value: &'a Value, name: &str) -> &'a Value {
-    value
-        .get(name)
-        .unwrap_or_else(|| panic!("JSON object should contain `{name}`"))
+    let Some(field) = value.get(name) else {
+        panic!("JSON object should contain `{name}`");
+    };
+    field
 }
 
 fn first_array_item(value: &Value) -> &Value {
-    value
-        .as_array()
-        .and_then(|items| items.first())
-        .expect("JSON value should be a non-empty array")
+    let Some(item) = value.as_array().and_then(|items| items.first()) else {
+        panic!("JSON value should be a non-empty array");
+    };
+    item
 }
 
 fn sample_agent_context() -> AgentContext {
