+++ ortho_config_macros/src/derive/parse/tests/mod.rs
@@ -3,4 +3,5 @@
 mod clap_attrs;
 mod lit;
 mod ortho_attrs;
+mod ortho_attrs_unknown_keys;
 mod type_utils;
